+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/errors/MiniAppBridgeError.kt
@@ -1,0 +1,11 @@
+package com.rakuten.tech.mobile.miniapp.errors
+
+/**
+ *  Contains the components to use custom errors from host app.
+ *  @property type The type of error.
+ *  @property message error message send to min app.
+ */
+open class MiniAppBridgeError(
+    type: String?,
+    message: String?
+)